requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
decorator==5.1.1
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-MySQLdb==1.0.1
Flask-WTF==1.0.0
idna==3.3
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
mysqlclient==2.1.0
phonenumbers==8.12.45
pycparser==2.21
PyJWT==1.4.2
PyMySQL==1.0.2
python-dotenv==0.20.0
six==1.16.0
validators==0.18.2
Werkzeug==2.0.3
WTForms==3.0.1
WTForms-Components==0.10.5
